Added nethercoal, fixed excavation and code clean
Problem: First, nethercoal was not an excavatable ore. Second,
excavation would not extend to ores of the same type if they were in
different stone. Third, the code for overriding the vanilla BTA blocks
was very messy, so it has been fixed.

Solution: Added nethercoal override, added test case for matching the
types of ores and cleaned up the code base.